% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorgrid.R
\name{colorgrid}
\alias{colorgrid}
\title{Plot TRAC Colors}
\usage{
colorgrid()
}
\value{

}
\description{
Creates a reference plot showing TRAC standard colors and their associated color name.
}
\examples{
colorgrid()
}
