% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcolor.R
\name{tcolor}
\alias{tcolor}
\title{Get TRAC Color}
\usage{
tcolor(tracID)
}
\arguments{
\item{tracID}{string}
}
\value{
string
}
\description{
Returns a string representing a TRAC standard colors in hex format.
}
\examples{
tcolor("gr9")

library(ggplot2)
ggplot() + geom_col(aes(x=c("foo", "bar"), y=1:2), fill = tcolor("bl1"))
}
