% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcolor.R
\name{tcolor}
\alias{tcolor}
\title{Returns TRAC standard colors in hex format}
\usage{
tcolor(tracID)
}
\arguments{
\item{tracID}{character}
}
\value{
character
}
\description{
Returns TRAC standard colors in hex format
}
\examples{
tcolor("trac_22")

ggplot() + geom_col(aes(x=x, y=value), color = tcolor("trac_22"))
}
